test: Try to re-work mail handling
Seems there's a race condition when more than one email is delivered,
and so the second time around, we get the previous email and the new
email.  This is an attempt to try and handle the buffering being done
by nodemailer/smtp-tester.

Tests at the moment are failing left right and centre.